requirements.txt
streamlit==1.28.0
openai==1.3.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.4.0
pandas==2.0.0
plotly==5.17.0
asyncio==3.4.3
httpx==0.25.0
PyYAML==6.0.1
pytest==7.4.0
python-multipart==0.0.6
